Format->guessFormat: Extend RDF/XML guessing
If the DOCTYPE element has to many ENTITY elements, the guess fails and returns null. That can happen, if there are many namespaces to define. With this patch, we extend the string-portion to check from 1024 to 3072 bytes of the content. In some cases that even may not enough, but at least, all major scenarios should be dealt with.